optimize code

+++ backend/model/src/account.rs
@@ -7,7 +7,7 @@ pub struct EvmAccountInfo {
     pub address: String,
     pub balance: String,
     pub balance_token: String,
-    pub nonce: u32,
+    pub nonce: u64,
     pub is_contract: bool,
     pub contract_type: Option<EvmContractTypeInfo>,
 }
@@ -16,6 +16,6 @@ pub struct EvmAccountInfo {
 pub struct EvmAccount {
     pub address: String,
     pub balance: String,
-    pub nonce: u32,
+    pub nonce: u64,
     pub is_contract: bool,
 }
