+++ backend/model/src/netwiork.rs
@@ -3,9 +3,9 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct EvmNetworkInfo {
     pub chain_id: u16,
-    pub gas_price: u32,
-    pub max_priority_fee: Option<u32>,
-    pub max_fee: Option<u32>,
-    pub latest_block_number: u32,
+    pub gas_price: u64,
+    pub max_priority_fee: Option<u64>,
+    pub max_fee: Option<u64>,
+    pub latest_block_number: u64,
     pub syncing: bool,
 }
